+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The following emojis are used to highlight certain changes:
 - `blockservice.NewWritethrough` deprecated function has been removed, instead you can do `blockservice.New(..., ..., WriteThrough())` like previously.
 - `gateway`: a new header configuration middleware has been added to replace the existing header configuration, which can be used more generically.
 - `namesys` now has a `WithMaxCacheTTL` option, which allows you to define a maximum TTL that will be used for caching IPNS entries.
+- `provider` the `ReproviderStats` now includes the time of the last and the next run of the reprovider.
 
 ### Changed